refactor(MovieListFragment): MVP 리팩터링
