+++ app/src/main/java/woowacourse/movie/view/moviemain/movielist/MovieListContract.kt
@@ -1,0 +1,9 @@
+package woowacourse.movie.view.moviemain.movielist
+
+import woowacourse.movie.view.model.MovieListModel
+
+interface MovieListContract {
+    interface Presenter {
+        fun getMovieListData(): List<MovieListModel>
+    }
+}
