+++ app/src/main/java/woowacourse/movie/view/moviemain/movielist/MovieListFragment.kt
@@ -7,19 +7,14 @@ import android.view.View
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.movie.R
-import woowacourse.movie.data.MovieMockRepository
-import woowacourse.movie.domain.Movie
-import woowacourse.movie.domain.repository.MovieRepository
 import woowacourse.movie.view.ReservationActivity
-import woowacourse.movie.view.mapper.toUiModel
 import woowacourse.movie.view.model.MovieListModel
 
 class MovieListFragment : Fragment(R.layout.fragment_movie_list) {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        val movieRepository: MovieRepository = MovieMockRepository
-        val movies = movieRepository.findAll()
-        val dataList = generateMovieListData(movies)
+        val presenter = MovieListPresenter()
+        val dataList = presenter.getMovieListData()
 
         val movieAdapter = MovieListAdapter(dataList, ::onClick)
         val movieListView = view.findViewById<RecyclerView>(R.id.movie_recyclerview)
@@ -39,34 +34,7 @@ class MovieListFragment : Fragment(R.layout.fragment_movie_list) {
         }
     }
 
-    private fun generateMovieListData(movies: List<Movie>): List<MovieListModel> {
-        val ad = MovieListModel.MovieAdModel(
-            R.drawable.woowacourse_banner,
-            "https://woowacourse.github.io/",
-        )
-
-        return mixMovieAdData(movies, ad, AD_POST_INTERVAL)
-    }
-
-    private fun mixMovieAdData(
-        movies: List<Movie>,
-        ad: MovieListModel.MovieAdModel,
-        adPostInterval: Int,
-    ): List<MovieListModel> {
-        val dataList = mutableListOf<MovieListModel>()
-        movies.forEachIndexed { index, movie ->
-            if (index % adPostInterval == adPostInterval - 1) {
-                dataList.add(movie.toUiModel())
-                dataList.add(ad)
-                return@forEachIndexed
-            }
-            dataList.add(movie.toUiModel())
-        }
-        return dataList
-    }
-
     companion object {
-        private const val AD_POST_INTERVAL = 3
         const val TAG_MOVIE_LIST = "MOVIE_LIST"
     }
 }
